tests: check define/undefine cycle
Make sure that defining and immediately undefining an interface actually
does remove the interface. This does not work because of a bug in Augeas
(https://fedorahosted.org/augeas/ticket/78)

The bug shouldn't affect us though, since we should reload Augeas on every
call through the public API to make sure we pick up changes to the
filesystem that happened in between calls.

  * tests/test-initscripts.c (testDefineUndefine): new test
  * tests/interface/bridge42.xml: interface for test