Move get_augeas and aug_submatch to dutil.c
In order to include augeas_xfm in the driver struct, the type of augeas_xfm
had to be reworked to a struct array. The inclusion of augeas_xfm in the
driver struct was necessary in order to move get_augeas and aug_submatch
from drv_initscripts.c to dutil.c. This also prepares for future utility
functions operating on the augeas instance.