Find interface config files according to initscripts' need_config
When we list interfaces or look them up, we need to account for the
possibility that there can be multiple ifcfg files with the same
DEVICE. For each such interface, we need to determine the 'preferred'
interface, using the same logic as in initscripts' need_config.

This is not a concern when undefining interfaces: we just unconditionally
nuke all the ifcfg files that have a matching DEVICE entry.

Part of the fix for BZ 512950