Fix initialization of libxslt
We used to initialize libxslt (and register extension modules) every time
ncf_init was called, and deregister/free libxslt globals every time
ncf_close was called. That breaks when libnetcf is used together with other
libraries that use libxslt, since they might continue using it after a call
to ncf_close.

To work around this, we now initialize and register our extension modules
every time we transform an XML doc, which obviates the need for global
cleanup (there's a small leak caused by global variables that libxslt
maintains, but that's negligible).

As a side-effect, we also capture error messages from the transform and
report them back with ncf_error.