ncf_if_xml_state: new API function
ncf_if_xml_state() returns the current live state of the interface, which
is just the same document returned by ncf_if_xml_desc(), with the
addition/modification of IP address and netmask (prefix) to the values
retrieved directly from the interface. This is useful to learn the IP
address assigned to an interface configured to used DHCP.

Any ip address retrieved in this manner will have an attribute "source" set
to "device", as a clue for the consumer.

Currently only the 1st IPv4 address on an interface is reported.