Remove all config-origined info from ncf_if_xml_state output.
Modify ncf_if_xml_state() to return *only* information queried from
the running device/kernel. Because this makes the "source='device'"
attribute redundant, also remove that.

Currently, we only query for the ipv4 address of the interface; if
there is none, don't even add a protocol node to the tree - just
return an "interface" document that has only a name, and nothing else.