Use netlink to get interface IP addresses.
While the SIOCGIFADDR ioctl can only get the first IPv4 address on an
interface, netlink reports all addresses, both IPv4 and IPv6 on the
interface.

The old function that uses the ioctl has been left in place, renamed,
to make the diffs simpler. It will be removed in the next patch.