Properly handle quoted entries in sysconfig files
The Augeas lens we used (Shellvars), maps an entry like

  DEVICE="eth0"

into the tree including quotes. That leads to problems down the road, since
the quotes become part of the interface name.

To address that, we use a lens (Sysconfig) that is more geared towards the
subset of shell syntax used in sysconfig files. It strips quotes when
reading files and adds and restores them as needed.

  * data/lenses/sysconfig.aug: lens for sysconfig files that strips
      quotes from values
  * data/lenses/tests/test_sysconfig.aug: tests for Sysconfig lens,
      especially for handling all the corner cases with quoted strings
  * src/drv_initscripts.c (augeas_xfm_common_pv): use Sysconfig lens,
      not Shellvars lens
  * Makefile.am: install Sysconfig lens
  * tests/root/etc/sysconfig/network-scripts/ifcfg-br0: add unnecessary
      quotes

Fixes BZ 613886