New APIs to implement transactional changes to network config
This patch provides three new public API functions that allow
reverting the network configuration back to a previous "known good"
state. The three new functions are:

  ncf_change_begin() - save a snapshot of current network config
  ncf_change_rollback() - revert to the config previously saved
  ncf_change_commit() - delete the saved snapshot, making the new
                        config permanent

If the system is rebooted after ncf_change_begin() is called, but
prior to calling ncf_change_commit(), the network config will be
automatically reverted to the saved state during the boot process.

The functionality is provided by the the script netcf-transaction,
which was added in the previous patch.

This initial version only updates the config files during a
rollback. It doesn't attempt to ifdown interfaces that are being
removed, ifup interfaces that are being added, or ifdown/ifup
interfaces that have been changed. That will be in an upcoming patch.