ncftool: Don't require "quit" command twice after a failed command
This is in response to: https://bugzilla.redhat.com/show_bug.cgi?id=681078

If an invalid command was entered at the ncftool prompt, and then
"quit" was immediately issued, it would be ignored - a 2nd "quit"
would be needed to exit from ncftool.

This was caused by main_loop() checking the wrong return status when
deciding to exit. Rather than checking the status of the command that
was just run, it was looking of the status of the command just prior
to that, and only exiting if that command was successful. What was
needed was to check the return status of the "quit" command (which
currently is always 0, but I suppose may change in the future.)