Remove unnecessary Requires: from netcf's pkgconfig file
This is in response to:

   https://bugzilla.redhat.com/show_bug.cgi?id=662056 (Fedora)
   https://bugzilla.redhat.com/show_bug.cgi?id=662057 (RHEL)

netcf.pc.in previously contained Requires for augeas, libxml, and
libxslt, but this was unnecessary - the pkgconfig file lists things
required to build an application that uses the netcf library, but
since netcf's use of those packages is purely internal/private, none
of those packages are actually required in order to build.

The solution is to remove everything from the Requires: line.