Capture stdout and stderr when execing external programs
This is in response to:

   https://bugzilla.redhat.com/show_bug.cgi?id=616060

but will also be useful in an upcoming patch which needs to retrieve
the output of /etc/rc.d/init.d/netcf-transaction.

Before forking, we now create a pipe, then after forking, the child
dup2's both stdout and stderr to the write side of the pipe, and the
parent calls fdopen() on the read side of the pipe and then calls
gnulib's fread_file() before calling waitpid() to reap the process.