exec_program: return unique exit codes for failure to exec or set sigmask
The child process can't do any reasonable error reporting about why it
failed, but it can return a different code to the parent for each
possible failure. Instead of returning 1 for all forms of failure
during exec, return 127 if the binary isn't found, 126 if it's
otherwise not executable, and 125 if pthread_sigmask() fails, then log
relevant errors in the parent.