add new NETCF_EINVALIDOP error code
NETCF_EINVALIDOP is the code to set when a particular function is
called while the system is in a state that doesn't allow it. This will
be used in the next patch.

While adding this new value to the enum, also #define all enums so
that a sourcefile can check for presence of any particular error code
at compile time with a simple #ifdef.