Don't log error if interface isn't found in kernel during status report
This patch resolves the following bug reports:

  https://bugzilla.redhat.com/show_bug.cgi?id=713286 (RHEL)
  https://bugzilla.redhat.com/show_bug.cgi?id=698108 (Fedora)

There are many situations that will result in the config files for an
interface being present, but the interface not being found in the
kernel's list of interfaces. For example, if a bridge interface has
been ifdowned, it will not exist at all. Another case is an "alias
interface" (eg "eth0:1", which is not really an interface on its own,
but just an obsolete method of specifying a 2nd (or more) IP address
for a single interface. (In the example given, the config files will
talk about the interface "eth0:1", but the kernel only knows about
"eth0".)

If we return an error status any time ncf_if_xml_state() can't find an
interface in the kernel, then management applications that
periodically poll netcf for the status of all system interfaces will
quickly collect scads of error log messages.

The solution is to ignore failure to find an interface in the kernel
(which in this situation is really just one more piece of information
about the interface's status), and just not fill in the information
that would have otherwise been gathered.