Always add <bridge> element to bridge if, even if no physdev is attached
This patch resolves the following bug filed against RHEL:

  https://bugzilla.redhat.com/show_bug.cgi?id=713180

If a bridge has no physical devices attached, ncf_if_xml_state() had
not been adding a <bridge> element to the returned XML. However,
<bridge> is required by the grammer if the interface is of
type="bridge" (even if it's empty), and libvirt enforces that.

The solution is to add in the <bridge> element *before* checking to
see if there are any physical devices attached.