Allow choice of network driver using --with-driver=XXXX
Instead of hardcoding the choice of driver, add a --with-driver=XXX
option accepting 'redhat' as the only choice for now. If the files
/etc/fedora-release or /etc/redhat-release are present, the 'redhat'
driver will be activated automatically. Otherwise configure will
exit with an error. The use of 'redhat' can be forced by specifiying
--with-driver=redhat, causing the /etc checks to be skipped

The existing --with-init-script=XXX option is now removed

The main library and test makefiles are re-organized so that
future addition of new drivers minimises the diff