support libnl-3 in addition to libnl-1
configure.ac:
If libnl-1 is not found, look for libnl-3, and (the also required)
libnl-route-3.

Makefile.am:
Add cflags and libs for libnl-3 and libnl-route-3

dutil_linux.{c,h}:
Switch the libnl calls over to new api if HAVE_LINBNL3.  Also do
not include <net/if.h> in that case, because <linux/if.h> is
included from the libnl3 include files, and can conflict with net/if.h.

Signed-off-by: Serge Hallyn <serge.hallyn@canonical.com>