eliminate "possible use before initialization" warning
In the case that aug_get_xml() errors out immediately, nmatches hasn't
been set, but is sent to free_matches() during cleanup. It turns out
that free_matches() only looks at nmatches if the 2nd arg is non-NULL
(which it won't be in this case), but initializing nmatches eliminates
the warning message.