Force --with-libnl1 for F17/RHEL6 and earlier
In order to assure that netcf on Fedora 17/RHEL6.x and earlier uses
libnl1 for its builds even if libnl3-devel is installed, the specfile
is changed to add --with-libnl1 to the configure line in those cases.

When libnl 3 support was added, the configure switch --enable-libnl1
was added, but that doesn't fit with the accepted standard usage of
--enable-* (which is supposed to enable a feature of the package being
built). Since it's new and hasn't been used until now, it's safe to
change it to the more standard -with-libnl1.