Fix bug that eats config parameters which have identical names.
Two (or more) nodes with identical names that are in the same
element are passed to libaugeas in such a way that only the last
node is actually written to the config file.
If e.g. there are two pre-up commands for a bridge on debian only
one is written to /etc/network/interfaces.
This fix changes this behaviour so that all of them are written.