Be sure to bail if get_augeas() fails
There were a few places where get_augeas() was called at the beginning
of a function, but the function was allowed to continue as normal even
if get_augeas() failed. This was disastrous if any augeas library
function was subsequently called prior to checking for error status -
get_augeas sets ncf->driver->augeas to NULL if there's an error, but
aug_get/aug_put/aug_rm (and possibly others) dereference the aug
pointer they are sent without checking for NULL.

This patch assures that any caller of get_augeas() returns immediately
if get_augeas() fails.